Rename pageStruture file
